Add files via upload
This code works completely, follows line for bar count 5, turns 90, goes into pocket, leaves the pocket, turns 90, follows line for bar count 1 and ends with a turn to heading to make sure we are straight. I added a negative distance for driving on heading. Also I added a negative distance for the helper function (getticks).